added transaction and exception handling to private_threads

+++ db/migrate/20091216084538_create_user_notifications.rb
@@ -1,7 +1,7 @@
 class CreateUserNotifications < ActiveRecord::Migration
   def self.up
     create_table :user_notifications do |t|
-      t.decimal :notification_id
+      t.integer :notification_id
       t.references :user
 
       t.timestamps
